Use 'TopLevel' names instead of temp names, remove duplicate docs
