+++ core/src/main/java/discord4j/core/object/entity/channel/Category.java
@@ -31,7 +31,7 @@ import java.util.Objects;
 import java.util.function.Consumer;
 
 /** A Discord category. */
-public final class Category extends BaseGuildChannelNotAThread {
+public final class Category extends BaseTopLevelGuildChannel {
 
     /**
      * Constructs an {@code Category} with an associated {@link GatewayDiscordClient} and Discord data.
