+++ core/src/main/java/discord4j/core/object/entity/channel/TopLevelGuildMessageChannel.java
@@ -1,0 +1,32 @@
+package discord4j.core.object.entity.channel;
+
+import discord4j.core.object.entity.Webhook;
+import discord4j.core.spec.WebhookCreateMono;
+import discord4j.core.spec.WebhookCreateSpec;
+import discord4j.core.spec.legacy.LegacyWebhookCreateSpec;
+import reactor.core.publisher.Flux;
+import reactor.core.publisher.Mono;
+
+import java.util.Optional;
+import java.util.function.Consumer;
+
+// TODO naming
+public interface TopLevelGuildMessageChannel extends CategorizableChannel, GuildMessageChannel {
+
+    @Override
+    Optional<String> getTopic();
+
+    @Override
+    Mono<Webhook> createWebhook(final Consumer<? super LegacyWebhookCreateSpec> spec);
+
+    @Override
+    default WebhookCreateMono createWebhook(String name) {
+        return GuildMessageChannel.super.createWebhook(name);
+    }
+
+    @Override
+    Mono<Webhook> createWebhook(WebhookCreateSpec spec);
+
+    @Override
+    Flux<Webhook> getWebhooks();
+}
